refactor: 임시저장 도메인관 관련된 컴포넌트 한 군데로 모음
