+++ frontend/src/components/tempArticle/TemporaryArticleItem/TemporaryArticleItem.styles.tsx
